vehicle-damage-assessment/requirements.txt
# core cv / ml
opencv-python>=4.8.0
ultralytics>=8.1.0        # YOLOv8
numpy>=1.24.0
scipy>=1.11.0
scikit-image>=0.21.0

# deep learning backend (pick one — torch is required by ultralytics)
torch>=2.0.0
torchvision>=0.15.0

# data / config
pyyaml>=6.0
pandas>=2.0.0

# visualization
matplotlib>=3.7.0

# utilities
tqdm>=4.65.0
Pillow>=10.0.0

# testing
pytest>=7.4.0
pytest-cov>=4.1.0

# optional: experiment tracking
# wandb>=0.15.0
# mlflow>=2.5.0
